code/requirements.txt
asgiref==3.8.1
async-timeout==5.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
dj-database-url==2.2.0
Django==4.2.16
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-redis==5.4.0
djangorestframework==3.15.2
h11==0.14.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kafka-python-ng==2.2.3
msgpack==1.1.0
packaging==24.1
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyOpenSSL==24.2.1
python-decouple==3.8
redis==5.2.0
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
tomli==2.0.2
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
uvicorn==0.32.0
whitenoise==6.7.0
zope.interface==7.1.1
